+++ app/src/renderer/vuex/getters.js
@@ -17,7 +17,8 @@ export const themes = state => state.themes
 export const onboarding = state => state.onboarding
 
 // wallet
-export const transactions = state => state.wallet.history
+export const allTransactions = state =>
+  state.transactions.wallet.concat(state.transactions.staking)
 export const wallet = state => state.wallet
 
 // staking