+++ app/src/renderer/vuex/modules/index.js
@@ -1,5 +1,6 @@
 export default opts => ({
   blockchain: require("./blockchain.js").default(opts),
+  transactions: require("./transactions.js").default(opts),
   config: require("./config.js").default(opts),
   delegates: require("./delegates.js").default(opts),
   delegation: require("./delegation.js").default(opts),