+++ test/unit/specs/store/wallet.spec.js
@@ -32,28 +32,10 @@ describe("Module: Wallet", () => {
     expect(store.state.wallet.balances).toEqual([])
   })
 
-  it("should set wallet history", () => {
-    const history = ["once", "upon", "a", "time"]
-    store.commit("setWalletHistory", history)
-    expect(store.state.wallet.history).toBe(history)
-  })
-
   it("should set denoms", () => {
     const denoms = ["acoin", "bcoin", "ccoin"]
     store.commit("setDenoms", denoms)
     expect(store.state.wallet.denoms).toBe(denoms)
-  })
-
-  it("should set transaction time", async () => {
-    const blockHeight = 31337
-    const time = 1234567890
-    const history = [{ height: blockHeight }]
-    const blockMetaInfo = { header: { time: time } }
-    store.state.blockchain.blockMetas = {
-      [blockHeight]: blockMetaInfo
-    }
-    let transactions = await store.dispatch("enrichTransactions", history)
-    expect(transactions[0].time).toBe(time)
   })
 
   // ACTIONS
@@ -67,7 +49,7 @@ describe("Module: Wallet", () => {
   it("should query wallet state", async () => {
     store.dispatch("queryWalletState")
     expect(store.state.wallet.balances).toEqual([])
-    expect(store.state.wallet.history).toEqual([])
+    expect(store.state.transactions.wallet).toEqual([])
     expect(store.state.send.nonce).toBe("0")
   })
 
@@ -96,41 +78,6 @@ describe("Module: Wallet", () => {
     expect(store.state.wallet.denoms).toEqual(["mycoin", "fermion", "gregcoin"])
   })
 
-  xit("should enrich transaction times", async () => {
-    node.coinTxs = () =>
-      Promise.resolve([
-        {
-          tx: {},
-          height: 1
-        },
-        {
-          tx: {},
-          height: 2
-        },
-        {
-          tx: {},
-          height: 2
-        }
-      ])
-    node.rpc.blockchain = ({ minHeight }, cb) => {
-      cb(null, {
-        block_metas: [
-          {
-            header: {
-              height: minHeight,
-              time: minHeight
-            }
-          }
-        ]
-      })
-    }
-    jest.spyOn(node.rpc, "blockchain")
-    await store.dispatch("queryWalletHistory")
-    expect(node.rpc.blockchain.mock.calls.length).toBe(2)
-    expect(store.state.wallet.history[0].time).toBe(1)
-    expect(store.state.wallet.history[1].time).toBe(2)
-  })
-
   it("should query the balances on reconnection", () => {
     store.state.node.stopConnecting = true
     store.state.wallet.balancesLoading = true
@@ -146,22 +93,6 @@ describe("Module: Wallet", () => {
     jest.spyOn(node, "queryAccount")
     store.dispatch("reconnected")
     expect(node.queryAccount).not.toHaveBeenCalled()
-  })
-
-  xit("should query the history on reconnection", () => {
-    store.state.node.stopConnecting = true
-    store.state.wallet.historyLoading = true
-    jest.spyOn(node, "coinTxs")
-    store.dispatch("reconnected")
-    expect(node.coinTxs).toHaveBeenCalled()
-  })
-
-  xit("should not query the history on reconnection if not stuck in loading", () => {
-    store.state.node.stopConnecting = true
-    store.state.wallet.historyLoading = false
-    jest.spyOn(node, "coinTxs")
-    store.dispatch("reconnected")
-    expect(node.coinTxs).not.toHaveBeenCalled()
   })
 
   it("should be in loading state before querying account, and not in loading state after", async () => {